+++ src/Libplanet.Net/Messages/IMessageCodec.cs
@@ -8,32 +8,17 @@ namespace Libplanet.Net.Messages
     {
         /// <summary>
         /// Encodes the message to <see typeref="T"/>-typed instance with given
-        /// <paramref name="privateKey"/> and <paramref name="peer"/>.
+        /// <paramref name="privateKey"/> for signing.
         /// </summary>
-        /// <param name="content">The message body to encode.</param>
+        /// <param name="message">The message to encode.</param>
         /// <param name="privateKey">The <see cref="PrivateKey"/> to sign the encoded message.
         /// </param>
-        /// <param name="appProtocolVersion">The <see cref="AppProtocolVersion"/> of
-        /// the transport layer.</param>
-        /// <param name="peer">The <see cref="BoundPeer"/>-typed representation of
-        /// the transport layer.
-        /// <seealso cref="ITransport.AsPeer"/></param>
-        /// <param name="timestamp">The <see cref="DateTimeOffset"/> of the time
-        /// <paramref name="content"/> is encoded.
-        /// </param>
-        /// <param name="identity">The byte array identifies the message to match between
-        /// message and its respond used in <see cref="NetMQ"/>.</param>
         /// <returns>A <see typeref="T"/> containing the signed <see cref="MessageContent"/>.
         /// </returns>
         /// <exception cref="InvalidCredentialException">Thrown when <paramref name="privateKey"/>'s
-        /// <see cref="PublicKey"/> does not match that of <paramref name="peer"/>.</exception>
-        T Encode(
-            MessageContent content,
-            PrivateKey privateKey,
-            AppProtocolVersion appProtocolVersion,
-            BoundPeer peer,
-            DateTimeOffset timestamp,
-            byte[]? identity);
+        /// <see cref="PublicKey"/> does not match that of <paramref name="message.Remote"/>.
+        /// </exception>
+        T Encode(Message message, PrivateKey privateKey);
 
         /// <summary>
         /// Decodes given <see typeref="T"/>-typed <paramref name="encoded"/> into